dom/base/test/fullscreen/test_fullscreen_modal.html

<!doctype html>
<title>Test for bug 1771150</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<style>
  #fullscreen {
    background-color: rgba(0, 255, 0, .5);
  }
  #fullscreen::backdrop {
    background-color: transparent;
  }
  #fullscreen, #fullscreen::backdrop {
    pointer-events: none;
  }
</style>
<div id="fullscreen"></div>
<button>Go fullscreen</button>
<script>
const button = document.querySelector("button");
let clickCount = 0;
let lastFullscreenPromise = null;
let shouldEnterFullscreen = false;
button.addEventListener("click", function() {
  clickCount++;
  if (shouldEnterFullscreen) {
    const fullscreenElement = document.getElementById("fullscreen");
    lastFullscreenPromise = SimpleTest.promiseFocus().then(() => {
      fullscreenElement.focus();
      return fullscreenElement.requestFullscreen();
    });
  }
});

function clickButton(expectEvent) {
  let lastClickCount = clickCount;
  synthesizeMouseAtCenter(button, {});
  (expectEvent ? isnot : is)(clickCount, lastClickCount, `Should've ${expectEvent ? "" : "not "}been able to click`);
}

function enterFullscreen() {
  lastFullscreenPromise = null;
  shouldEnterFullscreen = true;
  clickButton(true);
  shouldEnterFullscreen = false;
  isnot(lastFullscreenPromise, null, "Should be transitioning to fullscreen");
  return lastFullscreenPromise;
}

async function testFullscreenIsModal(modal) {
  info("testing modal: " + modal);
  is(document.fullscreenElement, null, "Shouldn't be in fullscreen");

  await enterFullscreen();

  clickButton(/* expectEvent = */ !modal);

  ok(document.fullscreenElement.matches(":fullscreen"), "Fullscreen element matches :fullscreen");
  is(document.fullscreenElement.matches(":modal"), modal, "Fullscreen element matches :modal");

  // Before exiting fullscreen, wait on the document becoming active,
  // which signifies that the fullscreen request has been completely
  // processed. It is important that the fullscreen request is all
  // the way done, or the exit fullscreen request could be dropped.
  let documentWrapper = SpecialPowers.wrap(document);
  await SimpleTest.promiseWaitForCondition(() => documentWrapper.isActive(),
                                           "Timed out waiting for document to become active.");
  ok(documentWrapper.isActive(), "Document is active.");

  await document.exitFullscreen();
  clickButton(/* expectEvent = */ true);
}

add_task(async function() {
  await testFullscreenIsModal(true);
});

</script>
